split out networking code from main business logic, add basic doc comments
